+++ .pylintrc
@@ -61,7 +61,7 @@ disable=fixme,
         old-style-class,
         superfluous-parens,
         bad-continuation,
-        unexpected-keyword-arg
+
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
