+++ .pylintrc37
@@ -148,8 +148,7 @@ disable=invalid-name,
         possibly-unused-variable,
         too-many-lines,
         too-many-statements,
-        bad-continuation,
-        unexpected-keyword-arg
+        bad-continuation
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
