+++ dash/exceptions.py
@@ -94,27 +94,3 @@ class SameInputOutputException(CallbackException):
 
 class MissingCallbackContextException(CallbackException):
     pass
-
-
-class DashTestingError(Exception):
-    """Base error for pytest-dash."""
-
-
-class InvalidDriverError(DashTestingError):
-    """An invalid selenium driver was specified."""
-
-
-class NoAppFoundError(DashTestingError):
-    """No `app` was found in the file."""
-
-
-class DashAppLoadingError(DashTestingError):
-    """The dash app failed to load"""
-
-
-class ServerCloseError(DashTestingError):
-    """The server cannot be closed"""
-
-
-class TestingTimeoutError(DashTestingError):
-    """"all timeout error about dash testing"""
